HomeFragment now updates list when switched to.
Added OnResume() for HomeFragment that updates the RecyclerView to add
any additional items to the view if there have been scans.

For some reason if you tab between the AddFragment and the HomeFragment
it does not update. However, if you tab between the SearchFragment and
the SettingsFragment it will update the list. Will look into this later.