Added DividerItemDecoration.java, updated HomeFragment and RecyclerViewAdapter
DividerItemDecoration.java adds horizontal bars between each item in the
RecyclerView.

The app was constantly re-adding items that were all ready populated in
the view causing duplicates. The ArrayList is now cleared upon starting
the fragment.

Updated RecyclerViewAdapter to properly displaying the room number. Had
to convert the integer to a string.