Edited list_item_inventory to use CardView and added a button.
I added a CardView to list_item_inventory so each item is surrounded by
a CardView similar to the CardView surrounding the UI in the
AddFragment. I also added a button to list_item_inventory so that each
item has a button when added to a RecyclerView. The goal is when this
button is pressed it passes the item fields to the add fragment to allow
for item editing.